Split out logger code (#778)
* Split out OutputChannelTransport and WinstonToolkitLogger

* Split out logger activation code; logger.ts now contains just the Logger interface and instance
