+++ src/shared/loggerUtils.ts
@@ -6,7 +6,8 @@
 import * as del from 'del'
 import * as path from 'path'
 import * as filesystemUtilities from './filesystemUtilities'
-import * as l from './logger/logger'
+import { initialize } from './logger/activation'
+import { WinstonToolkitLogger } from './logger/winstonToolkitLogger'
 
 export class TestLogger {
     /**
@@ -17,7 +18,7 @@ export class TestLogger {
      * @param logFolder - Folder to be managed by this object. Will be deleted on cleanup
      * @param logger - Logger to work with
      */
-    private constructor(private readonly logFolder: string, private readonly logger: l.WinstonToolkitLogger) {}
+    private constructor(private readonly logFolder: string, private readonly logger: WinstonToolkitLogger) {}
 
     public async cleanupLogger(): Promise<void> {
         this.logger.dispose()
@@ -32,14 +33,14 @@ export class TestLogger {
 
     public static async createTestLogger(): Promise<TestLogger> {
         const logFolder = await filesystemUtilities.makeTemporaryToolkitFolder()
-        const logger = await l.initialize({
+        const logger = await initialize({
             logPath: TestLogger.getLogPath(logFolder),
             logLevel: 'debug'
         })
 
         // In a future change, we will introduce a memory-based logger for testing usage.
         // For now, we rely on (and expect) a winston logger that is writing to a test log file.
-        if (logger instanceof l.WinstonToolkitLogger) {
+        if (logger instanceof WinstonToolkitLogger) {
             return new TestLogger(logFolder, logger)
         }
